% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{calls2stacktrace}
\alias{calls2stacktrace}
\title{Convert function call to stacktrace}
\usage{
calls2stacktrace(calls)
}
\arguments{
\item{calls}{function call}
}
\value{
a data.frame
}
\description{
Convert function call to stacktrace
}
