% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{calls_to_stacktrace}
\alias{calls_to_stacktrace}
\title{Convert function call to a stack trace}
\usage{
calls_to_stacktrace(calls)
}
\arguments{
\item{calls}{function calls, e.g. from sys.calls()}
}
\value{
a data.frame
}
\description{
Convert function call to a stack trace
}
\examples{
\dontrun{
f <- function() stop("cabin pressure lost")
f()
sentryR:::calls_to_stacktrace(sys.calls())
}
}
