% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{capture}
\alias{capture}
\title{Send a message to a Sentry server}
\usage{
capture(...)
}
\arguments{
\item{...}{named parameters}
}
\value{
sends message to Sentry
}
\description{
Send a message to a Sentry server
}
\examples{
\dontrun{
capture(message = "oh hai there!") # send message to sentry
}
}
