% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{capture_exception}
\alias{capture_exception}
\title{Report an error or exception object}
\usage{
capture_exception(error, ..., level = "error")
}
\arguments{
\item{error}{an error object}

\item{...}{optional additional named parameters}

\item{level}{the level of the message. Default: "error"}
}
\value{
nothing; sends error to Sentry
}
\description{
Report an error or exception object
}
\examples{
\dontrun{
capture_exception(simpleError("foo"), tags = list(version = "1.0"))
}
}
