% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_handler.R
\name{capture_function_calls}
\alias{capture_function_calls}
\title{Capture function calls}
\usage{
capture_function_calls(error)
}
\arguments{
\item{error}{error object}
}
\description{
Capture function calls
}
