% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentry.R
\name{capture_message}
\alias{capture_message}
\title{Report a message}
\usage{
capture_message(.message, ..., .level = "info")
}
\arguments{
\item{.message}{message text}

\item{...}{optional additional named paramters}

\item{.level}{the level of the message. Default: "info"}
}
\value{
nothing; sends message to Sentry
}
\description{
Report a message
}
\examples{
\dontrun{
capture_message("this is an important message", logger = "my.logger")
}
}
