% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentry.R
\name{configure_sentry}
\alias{configure_sentry}
\title{Configure Sentry}
\usage{
configure_sentry(dsn)
}
\arguments{
\item{dsn}{the DSN of a Sentry project.}
}
\value{
populates the .SentryEnv environment with character strings
}
\description{
Configure Sentry
}
\examples{
\dontrun{
configure_sentry("https://12345abcddbc45e49773bb1ca8d9c533@sentry.io/1234567")
.SentryEnv$host # sentry.io
}
}
