% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_handler.R
\name{default_error_handler}
\alias{default_error_handler}
\title{Default error handler for Plumber}
\usage{
default_error_handler(req, res, error)
}
\arguments{
\item{req}{a Plumber request object}

\item{res}{a Plumber response object}

\item{error}{an error object}
}
\value{
a list
}
\description{
Default error handler for Plumber
}
