% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{is_sentry_configured}
\alias{is_sentry_configured}
\title{Check if Sentry is configured}
\usage{
is_sentry_configured()
}
\value{
boolean
}
\description{
Check if Sentry is configured
}
\examples{
\dontrun{
configure_sentry("https://12345abcddbc45e49773bb1ca8d9c533@sentry.io/1234567")
is_sentry_configured() # TRUE
}
}
