% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{parse_dsn}
\alias{parse_dsn}
\title{Parse a Sentry DSN into its components}
\usage{
parse_dsn(dsn)
}
\arguments{
\item{dsn}{the DSN of a Sentry project.}
}
\value{
a named list with parsed elements of the DSN
}
\description{
Parse a Sentry DSN into its components
}
\examples{
parse_dsn("https://1234@sentry.io/1")
}
