% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentry.R
\name{prepare_payload}
\alias{prepare_payload}
\title{Prepare JSON payload for Sentry}
\usage{
prepare_payload(...)
}
\arguments{
\item{...}{named parameters}
}
\value{

}
\description{
Prepare JSON payload for Sentry
}
