% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentry.R
\name{sentry.config}
\alias{sentry.config}
\title{Parse Sentry DSN}
\usage{
sentry.config(dsn)
}
\arguments{
\item{dsn}{the DSN of a Sentry project as a character string.}
}
\value{
populates the .SentryEnv environment with character strings
}
\description{
Parse Sentry DSN
}
