% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_handler.R
\name{sentry_error_handler}
\alias{sentry_error_handler}
\title{Error handler with Sentry reporting}
\usage{
sentry_error_handler(req, res, error, ...)
}
\arguments{
\item{req}{a plumber request object}

\item{res}{a plumber response object}

\item{error}{an error object}

\item{...}{extra named variables for Sentry}
}
\value{
a list with response payload
}
\description{
Error handler with Sentry reporting
}
\examples{
\dontrun{
sentryR::configure_sentry(Sys.getenv("SENTRY_DSN"))
pr <- plumber::plumb("example_plumber.R")
pr$setErrorHandler(sentryR::sentry_error_handler)
pr$run()
}
}
