% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_handler.R
\name{with_captured_calls}
\alias{with_captured_calls}
\title{Create safe function}
\usage{
with_captured_calls(z)
}
\arguments{
\item{z}{the function whose errors we want to track}
}
\value{
a function
}
\description{
Create safe function
}
