% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{withCapturedCalls}
\alias{withCapturedCalls}
\title{Create safe function}
\usage{
withCapturedCalls(z)
}
\arguments{
\item{z}{the function whose errors we want to track}
}
\value{
a function
}
\description{
Create safe function
}
