% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_handler.R
\name{wrap_error_handler_with_sentry}
\alias{wrap_error_handler_with_sentry}
\title{Wrap a plumber error handler such that it reports errors to Sentry}
\usage{
wrap_error_handler_with_sentry(error_handler = default_error_handler)
}
\arguments{
\item{error_handler}{a function to handle plumber errors}
}
\value{
a function
}
\description{
Wrap a plumber error handler such that it reports errors to Sentry
}
