% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{capture_message}
\alias{capture_message}
\title{Report a message to Sentry}
\usage{
capture_message(message, ..., level = "info")
}
\arguments{
\item{message}{message text}

\item{...}{optional additional named parameters}

\item{level}{the level of the message. Default: "info"}
}
\value{
nothing; sends message to Sentry
}
\description{
Report a message to Sentry
}
\examples{
\dontrun{
capture_message("this is an important message", logger = "my.logger")
}
}
