% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentry.R
\name{configure_sentry}
\alias{configure_sentry}
\title{Configure Sentry}
\usage{
configure_sentry(dsn, appname = NULL, appversion = NULL)
}
\arguments{
\item{dsn}{the DSN of a Sentry project.}

\item{appname}{name of your application (optional). Default: NULL}

\item{appversion}{version of your application (optional). Default: NULL}
}
\value{
populates the sentry_env environment with character strings
}
\description{
Configure Sentry
}
\examples{
\dontrun{
configure_sentry("https://12345abcddbc45e49773bb1ca8d9c533@sentry.io/1234567")
sentry_env$host # sentry.io
}
}
